`Models::Vanilla`: allow specification of genes by name
Alternatively to the number of genes, the model can now be parametrized
by a vector of gene names. If this vector is specified, genes have to be
referred to by name in the link list. `SimulateScript` now also uses the gene
names for the column names if they are defined.

This commit also cleans up model specification parsing to make it more
robust and flexible.