Simulate: replace by `experiment` and rewrite specification logic
I completely rewrote how experiment specifications are parsed and
interpreted to allow identification of properties that are shared
between individual experiment items, such as model definition and
initial state. This is required for the upcoming `inspect` command to
understand how the data is structured.

The new specification language is preliminary, but the results structure
should be mostly final now. New and updated examples are included.

This commit also includes some minor changes to `GeneRegulatorySystems`
(mainly moving stuff around).