`Models.Vanilla`: remove volume/concentration considerations
We intend to eventually add this to a successor model, but in this
first iteration we consider the volume constant. To simplify the setup
we decided to fold the concentration calculation into the regulations'
half-activation parameters; it now needs to be done as part of the model
specification.

All examples were updated accordingly.