Add `Models.Kronecker`
This generates a `Models.Vanilla` for the given regulation network
initiators and parameters distributions. Setting a seed in the
specification makes the sampled model reproducible, so it can also be
recreated to be shown in the `InspectTool`.

The networks' link existence relation is sampled naively from a
stochastic Kronecker graph template; this is slow for larger networks,
but not as slow as simulating them. For now, the link aggregation for
the activation and repression networks is fixed to the minimum function.

To create networks at multiple scales but keep them comparable, it is
also possible to automatically adjust initiator matrices to achieve a
specified expected edge-to-node ratio in the expanded graph.
Specifically, all of the initiator's values are then offset by the same
constant on the logistic scale.