Experiment: include skipped segments in index
This is meant to simplify interpretation of the data (e.g. in
preparation for visualization) because this information no longer has to
be recreated from the `Schedule`'s `Specification`. Additionally, the
index now includes the number of slices associated with each segment,
and its numeric columns are now no longer `Dict`-encoded.

This also fixes a minor bug that caused the progress reporting
`DiscreteCallback` to emit stray trajectory slices.