Replace `Sysimage` functionality with PrecompileTools.jl
Although startup times are much slower that way (given already
precompiled images), this is less complex and brittle. Starting with
Julia v1.9, startup times for `Experiment` are now acceptable, given it
usually runs at least a couple of minutes for non-trivial workloads
anyway.

`Inspect` is still annoyingly slow to start, but it doesn't seem worth
it to pay the complexity cost just for this. At the moment, it is still
missing its `@compile_workload`, so maybe we can get it a little faster
(though it even so won't reach the startup times of using
PackageCompiler.jl).

The tools now each have their own Julia environment so that `Experiment`
and other potential future tools don't have to depend on Makie.jl.

This also raises the minimum Julia requirement to v1.9.