`Models.Vanilla`: make `polymerases` a proper species
The corresponding `Definition` entry now defines the name of the species
that should track the amount of polymerase; it is set to `"polymerases"`
by default and can be controlled e.g. with the `Models.Adjust` variants.
The name can also alias another species in the system, e.g. a gene's
proteins, which in this case will additionally take on the role of the
polymerase.