`Models.V1`: reaction overlay network refinements
Change the rate constant name to `rate` from `k`, and allow to
alternatively define `rates` to create a bidirectional reaction pair.

Also allow species references to either name a species directly, or to
name a gene, which is taken to mean the associated `proteins`. This
applies both to the regulation networks and to the reaction overlay
network and unifies these cases. Indirect (gene) references take
precedence.

Updating the dependencies also fixes the large stoichiometry issue which
was apparently a bug in Symbolics.jl triggered by upgrading to Julia
v1.10.