`Models`: more elaborate provenance tracking
`SciML.JumpModel` no longer contains a reference to the originating
definition. Instead, there is a new `Models.Derived` that abstracts this
tracking and which can be nested. This simplifies the construction of
model templates and also means that all intermediate definitions remain
available after model construction.

This also required an extension to the `Models.describe` mechanism, and
`InspectTool` will now display the full provenance chain in addition to
the regulation network.