`ReifyTool`: new tool to recreate objects from a Schedule specification
Given a results location and a path (e.g. as contained in the index'
path and model columns), this will reify the iteration state of the
scheduling mechanism at that point and can further index into the
resulting Schedule. In this way, ultimately, specific model definitions
can be obtained and dumped to standard output.

For "{regulation/v1}" and "{regulation/differentiation}" (for now), the
`--representation` flag can be used to convert the addressed object to a
nested `Dict`/`Array` data structure as it would be passed to the
`Specifications.constructor` mechanism. Using `--format=json`, this
object can then be written in the same JSON format that is used to
define models in the schedule specification language.