`Models.V1`: allow variable gene copy numbers
This renames the "promoter" species to "active", and allows specifying a
gene as `"unique": false`, which introduces a complementary "inactive"
species. In this case, the gene's copy number is given as the sum of
"active" and "inactive" and may vary dynamically. As a consequence,
"active" is then no longer binary, and one of these species has to be
added by intervention to make the copy number non-zero.

The default aggregation is reverted to `minimum` from `geometric_mean`.